+++ services/ui-src/src/measures/2023/shared/globalValidations/index.ts
@@ -4,7 +4,7 @@ export * as Types from "./types";
 
 export * from "shared/globalValidations/validateAtLeastOneDataSource";
 export * from "shared/globalValidations/validateAtLeastOneDefinitionOfPopulation";
-export * from "./validateAtLeastOneDataSourceType";
+export * from "shared/globalValidations/validateAtLeastOneDataSourceType";
 export * from "shared/globalValidations/validateAtLeastOneDeliverySystem";
 export * from "shared/globalValidations/validateDeviationTextFieldFilled";
 export * from "shared/globalValidations/validateAtLeastOneRateComplete";
