+++ services/ui-src/src/measures/2024/shared/globalValidations/index.ts
@@ -5,7 +5,7 @@ export * as Types from "./types";
 export * from "shared/globalValidations/validateAtLeastOneDataSource";
 export * from "shared/globalValidations/validateAtLeastOneDefinitionOfPopulation";
 export * from "shared/globalValidations/validateHybridMeasurePopulation";
-export * from "./validateAtLeastOneDataSourceType";
+export * from "shared/globalValidations/validateAtLeastOneDataSourceType";
 export * from "shared/globalValidations/validateAtLeastOneDeliverySystem";
 export * from "shared/globalValidations/validateDeviationTextFieldFilled";
 export * from "shared/globalValidations/validateAtLeastOneRateComplete";
